Update_Buttons.scss
Changed some discrepancies in terms of the naming convention.   &.rev-Button--inverted &.rev-Button--secondary were not following the same pattern established in the rest of the file- they were using color variables for background and vice versa. I added several new variables and utilized them in- &.rev-Button--inverted &.rev-Button--secondary - to solve this problem. Originally was tackling issue #395 but realized it was solved and merged after some time- realized the naming conventions were off in the process and wanted to protect the codebase against unexpected behavior / bugs.