typography update
potentially a breaking change. Instead of using the strip-unit function to remove the font-size unit and then add it back when we're setting the size, I updated it so that h1 gets the $h1 variable's font-size

I don't actually think this will break anything because you'll just the get value you assigned the header through the design tokens.

The function and accompanying math were to handle a potential issue with people adding px or pt values for font-size in the design tokens, instead of rem values.

I just tested adding px values for font-sizes and nothing bad happens